SVN_SILENT:
Merge r1156252 from e35 to trunk:

If we read an incidence from file that doesn't have an UID, call incidence->setUid( QString() ).

If we don't, then a random one is used, and on the second time we read from file (activate/deactivate resource)
we will have another random uid, so two incidences appear in the application.

Another solution would be to write the generated uid to file, but that's too soft and won't force people to fix
the application that generated faulty incidences.

Fixes kolab/issue3974

svn path=/trunk/KDE/kdepimlibs/; revision=1157401