Don't mix up CREATED and DTSTAMP, and introduce X-KDE-IMPLEMENTATION-VERSION.
CREATED is the creation-date of the icalendar object, and not the time of serialization.
DTSTAMP is the last-modification-date without METHOD, and the time of serialization with a METHOD.

This patch fixes the following problems:
* dtstamp and created exchanged when writing any ical object
* created date is not preserved
* last-modification-date is not preserved
* vFreebusy iTip messages contain a created date (but must not)

Backwardscompatibility is ensured by only applying the fix for ical
objects which have been written using the new implementation.

For this X-KDE-IMPLEMENTATION-VERSION has been introduced to track the
implementation version which makes it possible to provide proper
backwards compatiblity.

BUG: 310469
REVIEW: 107480
FIXED-IN: 4.9.5