Improve the documentation.
All incidences have an organizer. Even if it's just you watching
a movie alone, you're the organizer if that event.

pff, using shared pointers force you to be more explicit on what
can be returned. In KCal, this returned by value, so you knew it
wasn't null without reading the docs.