Work around KSystemTimeZones failing on the CI
Due to the dysfunctional KTimeZone daemon we never get a proper config
file, let alone in the right place for QStandardPaths test mode. So, just
write one ourselves, it can't be worse than not having one at all.

This is a temporary measure that becomes obsolete with completing the
QTimeZone port.