Port remaining API from KDateTime to QDateTime
The public interface of KCalCore is now free of KDateTime. Internally
we still use KTimeZone and KDateTime on a few places, those will be
ported in the next step.

We are no longer able to handle all custom timezones properly and we
cannot handle floating date-time properly (no ClockTime support). As
a result the Compat tests were adjusted to reflect the current state.
The only exception is the MSExchange test, for which we will add a
proper support in the next porting step.