Port from ICalTimeZone/KTimeZone to QTimeZone
A major disadvantage of QTimeZone compared to KTimeZone is that QTimeZone
does not allow us to specify custom time zones. This means that if we
parse an ICal file with custom non-standard time zones specified in the
VTIMEZONE section, we cannot just construct a QTimeZone from this data.

To workaround this limitation KCalCore will attempt to map the time zone
in VTIMEZONE to a nearest matching IANA time zone available in QTimeZone.
When writing back KCalCore will overwrite the original time zone speci-
fication with specification of the matched IANA time zone.