#include <strings.h> for strcasecmp in versit parser
Summary:
[[ http://pubs.opengroup.org/onlinepubs/9699919799/functions/strcasecmp.html | The POSIX definition for strcasecmp ]] states that it is in the <strings.h> include file, not <string.h>.  This causes the following build error on the musl libc:

> ```/usr/src/packages/user/kcalcore/src/kcalcore-17.08.1/src/versit/vcc.y: In function 'enterAttr':
> /usr/src/packages/user/kcalcore/src/kcalcore-17.08.1/src/versit/vcc.y:485:9: error: implicit definition of function 'strcasecmp' [-Werror=implicit-function-definition]```

This patch fixes this error (and a few others later on in the file).

Test Plan: Built and tested on glibc and musl Linux systems; since <strings.h> is a POSIX standard, it is assumed that other Unix-based systems (such as BSD and Solaris) provide it as well.

Reviewers: #kde_pim, winterz

Reviewed By: #kde_pim, winterz

Subscribers: winterz, #kde_pim

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D8255