Revert "Workaround qt5.13 bug"
This breaks the build on all platforms on the CI, and also causes all Dependency Builds to fail for Products which contain applications/libraries reliant on kcalcore.

Given the appropriate mailing list was informed, it appears that the state of this being broken is simply being ignored, which is not acceptable.
This commit may not be re-instated in any form without active consent from discussion on release-team.

CCMAIL: kde-pim@kde.org
CCMAIL: release-team@kde.org

This reverts commit f8d55fdd8ac27b27e7eb9e87a882f11a2aac5348.