Rename the KCalCore namespace to KCalendarCore
Summary:
As per discussion on kde-frameworks-devel. This is only the ABI-relevant
part so far, the API-relevant part in form of the include paths, as well
as documentation and various internal bits still need to be adjusted.

The ugly part here is the backward compatibility hack. An alias namespace
unfortunately isn't enough, it breaks with existing forward declarations.
We therefore need a way to rename the old KCalCore name in code that does
not include any of our headers too. All KDE PIM code builds unmodified
with this.

Subscribers: kde-pim

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D22697