Update visibility cache when notebook visibility is changed
Summary:
incidence visibility is cached in mIncidenceVisibility hash array. But this array is not updated when a notebook visibility is changed when calling updateNotebook().

It also adds a method to inquire the visibility of a notebook. Previously, one can only inquire the visibility of an incidence, while the visibility flag is set for a notebook.

Test Plan: add a test in testmemorycalendar.cpp to demonstrate the issue and the fix.

Reviewers: vkrause, dvratil, kde-pim, winterz

Reviewed By: winterz

Subscribers: #kde_pim

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D27826