Fix serialization of recurring to-do's start date
KCalendarCore uses Todo::Private::mDtRecurrence to hold the start date
and time of the next uncompleted occurrance of a recurring Todo, and
deserializes it from the X-KDE-LIBKCAL-DTRECURRENCE property of iCalendar
files.  However, ICalFormatImpl::writeTodo() serializes dtDue() into that
property, not dtStart().

When KOrganizer creates a new recurring to-do, it serializes the to-do,
then deserializes it to populate the To-Do List view, so the view shows
an incorrect start date (and an incorrect due date, which is calculated
from the start date).

BUG: 345498
BUG: 345565