Read floating date time as LocalTime
Date time like DTSTART:20191113T120000 are considered like floating
date time in RFC5545. This commit modifies the readICalDateTime
function not to use the system time zone in that case, but the
LocalTime spec. Otherwise, there is no way to distinguish incoming
floating time from time given in the system time zone. Then when
exporting to iCal format, the floating format is not conserved.

Also properly export date time in UTC using the Z character and not TZID=UTC
as mentioned in RFC5545 section 3.3.5:
"The "TZID" property parameter MUST NOT be applied to DATE-TIME
 properties whose time values are specified in UTC."