Update Compat test references
Reorder incidences and ensure that QHash storing them will always
use a predictable ordering in the tests.
Remove the spurious TZID=UTC to use the 'Z' notation instead.
Correct some date and time that were incorrect in the reference
when compared to the original source.
Use the GMT time zone instead of the GM one that is not defined.