Implement PERIOD support in RDATE
PERIOD is defined in the RFC5545, see section-3.8.5.2 and can be used instead
of exceptions to define recurring occurrences with a different duration than
the parent incidence.

Also add OccurrenceIterator::occurrenceEndDate() to hide the period internal
implementation and provide a high level routine to get the ending date time
if it makes sense for the incidence type.

CCBUG: 271308