Fix timezone information being lost when creating events
setDtStart() was doing nothing even though the timeSpec was changed.

QDateTime::operator== sees the same moment in time when comparing
"12:00 Qt::TimeZone Europe/Paris" and "12:00 LocalTime", so it
returns true.

However in kcalendarcore, very likely since
commit 670ebe21315e0440, Qt::LocalTime means a floating event,
which isn't the same thing as an event with a timezone.

The resulting ICAL serialization was missing timezone
information completely.