Compare QDateTime times, timespecs, and timezones
Add a function to compare two QDateTimes to see if their timeSpecs and
time zones are equal, in addition to the moment in time that they
represent.  Use it to decide whether incidences have been dirtied,
or are equal.

I declared the new `identical()` function in `incidencebase.h` because I 
plan to call it in the incidence editor, and didn't see a better choice
of include file.

I call it in `Period::operator==()`, `Recurrence::Private::operator==()`,
and `RecurrenceRule::Private::operator==()` because as I read the RFC
those datetimes can have time zones or be floating.