Add calendar loading state property
This is useful to indicate an asynchronously loaded calendar is still being
filled and thus read access will not return complete (or any) results.

This is something the various Calendar sub-classes in akonadi-calendar
use for example, upstreaming this here will allow to check for this
generically when using the platform calendar plugin interface as well.