Deprecate the notebook support
Notebooks provide a way to partition a calendar into different subsets,
or the other way around, build up an aggregate calendar from different
sources. This was originally contributed as result of Nokia's Maemo
work.

This isn't something we use in KDE (we do this the other way around),
and arguably this is the wrong layer to have such functionality on
anyway (it should rather be a layer on top of KCalCore::Calendar).

It does have a large API surface and complicates Calendar and CalFormat
internals even further, so let's use the opportunity of KF6 to drop it.