Deprecate Calendar relations API
This is unused, and there's a second implementation of this in
Akonadi::CalendarBase which is actually used. We can't move that
here as that leverages the more fine-grained change notifications
available there.

Design-wise, this is also not something that belongs into Calendar
itself in the first place, but should rather be a thing living alongside
a Calendar instance. It's also unecessarily restrictive by implying a
specific type of relation, while e.g. Zanshin has multiple types of
relations.