MemoryCalendar: log details about duplicate events before asserting
When trying to insert an incidence into a MemoryCalendar that has a UID
that is already present in the calendar, first log as much as possible
before asserting that it's the same insert being re-inserted.

This does not fix anything, but might help with diagnostics, and provides
some useful and actionable output before asserting.

CCBUG: 481031