Fix check whether RRULE UNTIL datetime is UTC
In Qt6, the timeSpec has changed behavior, and just checking whether
a QDateTime::timeSpec() == Qt::UTC is not enough, since when the
QDateTime has a valid QTimeZone (that represents UTC), timeSpec() will
return Qt::TimeZone.

Thus the proper way is to check whether the QTimeZone returned by QDateTime
is the UTC timezone.

This fixes "UNTIL in RRULE must be specified in UTC" warning printed by
the testicaltimezones test.