Fix conversion of date-only icaltimetype to UTC QDateTime
The code in readICalDateTime converts the values in icaltimetype structure
to a QDateTime, using the timezone specified in the icaltimetype. If the
icaltimetype is day-only (e.g. in all-day incidence) then it does not
have any timezone information and so Qt::LocalTime is assumed, and a new
QDateTime is constructed that happens on the midnight (beginning) of the
specified day, in user's current local timezone.

In some cases, the function is asked to return the QDateTime in UTC timezone
(for fields that the RFC mandates be always in UTC, like RRULE's UNTIL).
If the user's local timezone is in positive UTC offset, the QDateTime
gets converted to UTC, which shifts those \"all-day\" QDateTimes back
by one day. Later in the code, when it realizes it should be dealing only
with a date, the timezone information is discarded and we end up with a
QDate(Time) that is off by one day compared to what's stored in the iCal data.

This fixes simplifies the handling of day-only icaltimetype and makes sure
to construct the QDateTime already in the correct timezone (i.e. UTC) so
no more conversions are necessary.

BUG: 483707
FIXED-IN: 6.1.0