Avoid computing the next recurrence interval based on an invalid time
When Qt is build with asserts enabled this will abort otherwise, as seen
in the FreeBSD CI.

In the unit tests we seem to hit this for recurring todos without a
starting point, which doesn't seem entirely invalid. In those cases take
our best guess for a starting point and continue from there, which for
things like daily recurring tasks should be fine.

Alternative attempts with rejecting such recurrences outright break tests,
so I assume that is not an option.