Add read support for xCal events
This is not exposed as public API (yet) on its own as it's still not
entirely complete. It's only accessible via FileStorage, allowing that
to read more content than before.

In particular this only supports reading and supports events so far,
but for those both the final RFC 6321 variant as well as an older draft
standard.

This is motivated by Pretalx having removed iCal export support, something
Kongress relied on. Pretalx still has xCal export though (using the older
RFC draft).