Repair test on supported FreeBSD versions
An older workaround that expected differences in Prague's timezone
transitions between modern Linux and FreeBSD-at-the-time (and also
old Debian) is no longer necessary. FreeBSD 13.2 yields the same
timezone transitions for Prague as expected on Linux.

I can't tell if this is a difference in the tzinfo input, or changes
to zic, but removing the special-cases for FreeBSD repairs the test.