Update recurrenceId when dtStart changes
When changing the dtStart of a recurring event,
the recurrence ids of exceptions don't match
anymore any of the occurrence dtStart. Exceptions
are thus ignored from the instance iterator, but
they are still listed as exceptions by the
Calendar::instances() method.

Correct this inconsistency by shifting the
recurrence ids of exceptions accordingly with
the change in dtStart of the recurring incidence.